Time: 1 ms (93%), Space: 42.7 MB (31.89%) - LeetHub

+++ 0029-divide-two-integers/0029-divide-two-integers.java
@@ -3,6 +3,6 @@ class Solution {
         if(dividend==Integer.MIN_VALUE && divisor==-1){
             return Integer.MAX_VALUE;
         }
-        return (int) dividend/divisor;
+        return dividend/divisor;
     }
 }
